pyproject.toml
[tool.poetry]
name = "demetriusstorm-gendiff"
version = "0.5.0"
description = ""
authors = ["Dmitriy Bulygin <demetrius.storm@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^5.4.3"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.engine:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
